testing/requirements.txt
pyjwt[crypto]
requests